pyproject.toml
[project]
name = "dart-fss-text"
version = "0.1.0"
description = ""
authors = [
    {name = "jaepil-choi.laptop",email = "chljeffreyz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "dart-fss (>=0.4.14,<0.5.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]
packages = [{include = "dart_fss_text", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "-m", "not smoke"  # Skip smoke tests by default
]
markers = [
  "unit: fast, no network",
  "integration: may touch filesystem/network",
  "smoke: quick live API sanity checks (requires API key, disabled by default)"
]